requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
async-timeout==5.0.1
attrs==25.4.0
backoff==2.2.1
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
defusedxml==0.7.1
exceptiongroup==1.3.1
fastapi==0.123.8
fonttools==4.60.1
frozenlist==1.8.0
h11==0.16.0
idna==3.11
importlib_resources==6.5.2
inference-sdk==0.62.2
kiwisolver==1.4.7
marshmallow==3.26.1
matplotlib==3.9.4
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.0.2
opencv-python==4.10.0.84
packaging==25.0
pillow==11.3.0
propcache==0.4.1
py-cpuinfo==9.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.5
scipy==1.13.1
six==1.17.0
starlette==0.49.3
supervision==0.27.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
yarl==1.22.0
zipp==3.23.0
